put pdf button below the table and change text
