+++ frontend/src/locales/messages.en.json
@@ -28,6 +28,7 @@
   "payments.expenseSheetsReadyForPaymentTable.name": "Name",
   "payments.expenseSheetsReadyForPaymentTable.no_expenses_ready_for_payment": "No expenses ready for payment.",
   "payments.expenseSheetsReadyForPaymentTable.start_payment": "Start payment",
+  "payments.expenseSheetsReadyForPaymentTable.pdf": "Export PDF",
   "payments.expenseSheetsReadyForPaymentTable.zdp": "ZDP",
   "payments.paymentDetail.amount": "Amount",
   "payments.paymentDetail.cancel_payment": "{icon} Cancel payment",
@@ -113,7 +114,7 @@
   "store.expenseSheetStore.expense_sick_days_not_loaded": "Sick days could not be loaded",
   "store.expenseSheetStore.expense_sheet.one": "The expense sheet",
   "store.expenseSheetStore.expense_sheet.other": "The expense sheets",
-    "store.expenseSheetStore.expense_sheet.no_user_dime": "No user could be found in Dime",
+  "store.expenseSheetStore.expense_sheet.no_user_dime": "No user could be found in Dime",
   "store.expenseSheetStore.expense_sheet_per_site": "expense sheets per page:",
   "store.expenseSheetStore.state_changed": "{entityNameSingular} was changed to {state}.",
   "store.holidayStore.holiday.one": "The holiday",
